# đźš€ Add optimized startup scripts and automatic log cleanup