… LOG_CLEANUP_FEATURE.md
… OPTIMIZATION_SUMMARY.md
… OPTIMIZED_STARTUP_GUIDE.md
… setup-environment.bat
… start-dev-optimized.bat
… start-prod-optimized.bat
… start-universal.bat
… switch-config.bat
… test-log-cleanup.bat